Close the release gate, pin the last three deps, and stop truncating the clock
Four findings from the review, all in production code rather than tests.

THE RELEASE GATE HAD NO GATE. check_version.sh recognised a release tag as /^v[0-9]/
while release.yml triggers on the far wider "v*", so a tag of vtest, v-wip or a bare v
left $tag empty, printed the declared version and exited 0 -- a release built and
published with no version check at all. release.yml names its release ${REF_NAME#v},
so vtest would have shipped as "test".

The discriminator is now the ref TYPE rather than the shape of the name, which is what
it always meant: GITHUB_REF_TYPE=tag. A branch is never read as a version whatever it
is called, which is the false positive the old name-shape guard existed to avoid.

It also gets a --self-test, run in CI on every push rather than only at tag time -- a
broken gate should surface before the release that needs it. The first version of that
self-test was itself vacuous: it passed each tag as an ARGUMENT, and every case still
passed with the bug reintroduced, because the argument path was never the broken one.
It now drives the environment with no argument, exactly as release.yml does, and the
three tags I found empirically are the three it fails on.

DEPENDENCIES. ArduinoJson and espMqttClient carried carets (^7.4.3, ^1.7.3), which is
not a pin: it accepts any later minor. .pio/ is gitignored and there is no lockfile, so
the resolved set lived only in whatever machine built last -- and this builds an OTA
image, where "the version that shipped" has to be reconstructible from the tag. Both
were already resolving to the versions now named, and both are the latest release
upstream, so this pins current behaviour rather than changing it.

THE CLOCK. main.cpp narrowed nowMs() to uint32 on the way into breadcrumbs::tick(),
forty lines below a comment warning that millis() wraps every 49.7 days and that
casting a wrapped value to uint64 does not un-wrap it. The record exists to say "this
life reached this uptime", so the bridges it was built for -- the long-lived ones --
were exactly the ones it misreported, and it misreported them as a plausible smaller
number: a 60-day life read as 10.3 days.

Storage now holds SECONDS in the same uint32, which reaches 136 years and keeps the
struct at sixteen bytes, so the CRC window and the RTC-RAM layout are untouched.
Nothing is lost: tick() only ever wrote once per second.

  EXTERNAL BEHAVIOUR CHANGE, flagged deliberately. /api/v1/diagnostics still publishes
  previous_uptime_ms in milliseconds, but the value is now always a whole number of
  seconds. Accuracy is unchanged -- it was already only sampled once per second -- but
  a consumer reading the low digits will see zeros where it used to see noise. The
  field also now carries values past 2^32, and the payload is tested for that.

MQTT THROTTLE. PublishThrottle::Sample held its id as a std::string, so every publish
rebuilt the whole vector: one allocation per channel, freed on the next, at least once
a minute per device for the life of the bridge. measurement.h keeps ids as const char*
for exactly that reason -- it records 80-120 short-lived allocations per poll as "a
classic fragmentation hazard" -- and this was the one place downstream that put them
back on the heap. Every id is an inline constexpr constant with static storage, so the
pointer is safe to borrow. The comparison is now an explicit strcmp: it would have
silently become a pointer comparison otherwise, which happens to work for interned
constants and is not something to depend on.

Verified: 1024 native cases, check_layering.sh, the new self-test, and a real
waveshare-rs485-can build with no new warnings.

+++ src/device/bridge_info.h
@@ -113,7 +113,7 @@ struct BridgeInfo {
     /// reason is not here: it is this boot's resetReason, already above.
     uint32_t bootCount        = 0;
     bool     breadcrumbsCold  = true;
-    uint32_t previousUptimeMs = 0;
+    uint64_t previousUptimeMs = 0;
     uint32_t previousFirmware = 0;
 
     /// The board this firmware is running on. Reported to Home Assistant as the bridge
